Create DesignaStoreDatabase.KhanAcademy1
I am creating my store! This store sells different bikes. It has a table for all the store items and at least 5 columns for the data that needs to be stored. This store sells at least 15 items. A select statement is used to order the items by price. I also performed statistics.